Added export to CSV functionality to deck registration page
