+++ app/frontend/players/Players.svelte
@@ -25,6 +25,7 @@
     loadIdentityNames,
     type IdentityNames,
   } from "../identities/Identity";
+    import { deckCsv } from "../utils/decks";
 
   let { tournamentId }: { tournamentId: number } = $props();
 
@@ -171,53 +172,10 @@
     }
 
     const decks = await loadDecks(tournamentId);
-    const headerCsv =
-      decks
-        .map((deck) => `Player,${quoteCsvValue(deck.details.player_name)},`)
-        .join(",,") +
-      "\n" +
-      decks
-        .map((deck) => `Deck,${quoteCsvValue(deck.details.name ?? "")},`)
-        .join(",,") +
-      "\n\n" +
-      decks.map(() => "Min,Identity,Max").join(",,") +
-      "\n" +
-      decks
-        .map(
-          (deck) =>
-            `${deck.details.min_deck_size},${quoteCsvValue(deck.details.identity_title)},${deck.details.max_influence}`,
-        )
-        .join(",,");
-
-    const maxCards = decks.reduce(
-      (max, deck) => Math.max(max, deck.cards.length),
-      0,
-    );
-    let cardCsv = decks.map(() => "Qty,Card Name,Inf").join(",,");
-    for (const i of Array(maxCards).keys()) {
-      cardCsv +=
-        "\n" +
-        decks
-          .map((deck) =>
-            i < deck.cards.length
-              ? `${deck.cards[i].quantity},${quoteCsvValue(deck.cards[i].title)},${deck.cards[i].influence > 0 ? deck.cards[i].influence : ""}`
-              : ",,",
-          )
-          .join(",,");
-    }
-    cardCsv +=
-      "\n\n" +
-      decks
-        .map(
-          (deck) =>
-            `${deck.cards.reduce((total, card) => total + card.quantity, 0)},Totals,${deck.cards.reduce((total, card) => total + card.influence, 0)}`,
-        )
-        .join(",,");
-
     downloadBlob(
       `Decks for ${data.tournament.name}.csv`,
-      new Blob([`\ufeff${headerCsv}\n\n${cardCsv}`], { type: "text/csv" }),
-    ); // "\ufeff" lets Excel know it's Unicode encoded
+      new Blob([deckCsv(decks)], { type: "text/csv" }),
+    );
   }
 
   function downloadStreamingSpreadsheet() {
